Connect DB to remote instance, Route reconfig
Moved database connection variables to environment variables.
They need to be set in a .env file, in the same directory the server is
running from (ideally in the '/server' directory).

Cleaned up readme and unneeded sqlite file.

Added route /to test operations with the new database. It will
attempt to connect, create a string with id, push it, fetch it back, and
return it as a response to the frontend.

Reconfigured the routes to hopefully make more sense.
- On index route, generic Hello World. Returns a static string from
  server.
- On /home route, Hello World from the database (fullstack). Utilizes the
  Django ORM and cloud database to create, update, and fetch data. Then
  returns the fetched data as a response.